pyproject.toml
[project]
name = "skweezy-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "gtts>=2.5.4",
    "langchain-community>=0.4.1",
    "langchain-core>=0.3.0",
    "langchain-groq>=1.1.2",
    "langchain-text-splitters>=0.3.0",
    "langchain-classic>=0.3.0",
    "chromadb>=1.0.0",
    "llvmlite==0.43.0",
    "onnxruntime==1.16.3",
    "pypdf2>=3.0.1",
    "streamlit>=1.55.0",
    "torch==2.2.2",
    "transformers>=5.4.0",
    "sentence-transformers>=3.0.0",
    "Pillow>=10.0.0",
    "pandas>=2.0.0",
    "unstructured>=0.10.0",
    "python-dotenv>=1.0.0",
    "validators>=0.35.0",
    "yt-dlp==2024.7.9",
]
